+++ test/checks_mpi.sh
@@ -132,7 +132,7 @@ done
 rep=OrszagTang3D
 cd $TEST_DIR/MHD/$rep
 # remove generated vtk from previous run
-rm *.vtk
+mv data.0001.vtk data.0001.old.vtk
 echo "***********************************************"
 echo "Running  $rep with restart dump # 1"
 echo "***********************************************"
@@ -143,3 +143,8 @@ echo "Testing  $rep with restart dump # 1"
 echo "***********************************************"
 python3 testidefix.py -noplot || { echo "!!!! MHD $rep failed checking restart dump validation"; exit 1; }
 cd ..
+echo "***********************************************"
+echo "Checking bitwise compatibility of output from restarts"
+echo "***********************************************"
+diff data.0001.vtk data.0001.old.vtk || { echo "!!!! MHD $rep failed: restart dumps do not produce exactly the same results"; exit 1; }
+echo "Success"